Move _app to top level module
