+++ main/hooks.js
