+++ main/index.jsx
@@ -1,0 +1,4 @@
+// TODO: Create App Layout
+const Main = ({ children }) => children
+
+export default Main
